refactor: move compute_payload_hash from router to service layer
Follows project layering pattern: routers → services → models.
The function now lives in pow_service.py alongside related PoW functions.

Closes #93
